+++ src/user.ts
@@ -10,6 +10,7 @@ export type User = {
   denyPermission(descriptor: PermissionDescriptor): void;
   resetPermission(descriptor: PermissionDescriptor): void;
   requestAccess(descriptor: PermissionDescriptor): Promise<boolean>;
+  setAccessRequestHandler(nextHandler: HandleAccessRequest): void;
 };
 
 export function createUser({
@@ -63,6 +64,10 @@ export function createUser({
       }
 
       return shouldAllow;
+    },
+
+    setAccessRequestHandler(nextHandler) {
+      handleAccessRequest = nextHandler;
     },
   };